# backend/accounts-backend/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Accounts Backend API
  version: 1.0.0
  description: API REST para clientes, cuentas, movimientos y reportes
  contact:
    name: Jefferson Noroña

servers:
  - url: http://localhost:8081

tags:
  - name: Clientes
  - name: Cuentas
  - name: Movimientos
  - name: Reportes

paths:
  /api/clientes:
    get:
      tags: [Clientes]
      summary: Listar clientes
      operationId: listClientes
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - in: query
          name: q
          schema:
            type: string
          description: Búsqueda rápida
      responses:
        '200':
          description: Página de clientes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponseCustomer'
        '400':
          $ref: '#/components/responses/BadRequest'
    post:
      tags: [Clientes]
      summary: Crear cliente
      operationId: createCliente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerCreate'
      responses:
        '201':
          description: Cliente creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'

  /api/clientes/{clienteId}:
    get:
      tags: [Clientes]
      summary: Obtener cliente
      operationId: getCliente
      parameters:
        - $ref: '#/components/parameters/ClienteIdParam'
      responses:
        '200':
          description: Cliente encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerDTO'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Clientes]
      summary: Actualizar cliente
      operationId: updateCliente
      parameters:
        - $ref: '#/components/parameters/ClienteIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerUpdate'
      responses:
        '200':
          description: Cliente actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Clientes]
      summary: Actualizar parcialmente cliente
      operationId: patchCliente
      parameters:
        - $ref: '#/components/parameters/ClienteIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerPatch'
      responses:
        '200':
          description: Cliente actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Clientes]
      summary: Eliminar cliente
      operationId: deleteCliente
      parameters:
        - $ref: '#/components/parameters/ClienteIdParam'
      responses:
        '204':
          description: Cliente eliminado
        '404':
          $ref: '#/components/responses/NotFound'

  /api/cuentas:
    get:
      tags: [Cuentas]
      summary: Listar cuentas
      operationId: listCuentas
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - in: query
          name: clienteId
          schema:
            type: string
            format: uuid
        - in: query
          name: q
          schema:
            type: string
      responses:
        '200':
          description: Página de cuentas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponseAccount'
        '400':
          $ref: '#/components/responses/BadRequest'
    post:
      tags: [Cuentas]
      summary: Crear cuenta
      operationId: createCuenta
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountCreate'
      responses:
        '201':
          description: Cuenta creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'

  /api/cuentas/{cuentaId}:
    get:
      tags: [Cuentas]
      summary: Obtener cuenta
      operationId: getCuenta
      parameters:
        - $ref: '#/components/parameters/CuentaIdParam'
      responses:
        '200':
          description: Cuenta encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountDTO'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Cuentas]
      summary: Actualizar parcialmente cuenta
      operationId: patchCuenta
      parameters:
        - $ref: '#/components/parameters/CuentaIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountPatch'
      responses:
        '200':
          description: Cuenta actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Cuentas]
      summary: Eliminar cuenta
      operationId: deleteCuenta
      parameters:
        - $ref: '#/components/parameters/CuentaIdParam'
      responses:
        '204':
          description: Cuenta eliminada
        '404':
          $ref: '#/components/responses/NotFound'

  /api/movimientos:
    get:
      tags: [Movimientos]
      summary: Listar movimientos
      operationId: listMovimientos
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - in: query
          name: cuentaId
          schema:
            type: string
            format: uuid
        - in: query
          name: from
          schema:
            type: string
            format: date
        - in: query
          name: to
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Página de movimientos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponseTransaction'
        '400':
          $ref: '#/components/responses/BadRequest'
    post:
      tags: [Movimientos]
      summary: Registrar movimiento bancario
      description: |
        Registra un movimiento bancario (crédito o débito) aplicando reglas de negocio.
        
        REGLAS DE NEGOCIO:
        1. CREDITO: Incrementa el saldo (valores positivos)
        2. DEBITO: Decrementa el saldo (valores negativos)
        3. Límite diario de retiro: $1000 USD ACUMULADO por día
           - Se suma TODOS los débitos del día (00:00-23:59)
           - Si acumulado_débitos_día + nuevo_débito > $1000, se rechaza con código 1002
           - Ejemplo: Retiro $575 + Retiro $426 = $1001 > $1000 (RECHAZADO)
           - Ejemplo: Retiro $575 + Retiro $425 = $1000 (PERMITIDO)
        4. Saldo cero: Si saldo = 0, no se permite débito (código 1001)
        5. Saldo insuficiente: Si saldo < monto, no se permite (código 1001)
      operationId: createMovimiento
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionCreate'
      responses:
        '201':
          description: Movimiento registrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      tags: [Movimientos]
      summary: Anular movimiento
      operationId: deleteMovimiento
      parameters:
        - in: query
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Movimiento anulado
        '404':
          $ref: '#/components/responses/NotFound'

  /api/reportes:
    get:
      tags: [Reportes]
      summary: Generar reporte
      operationId: getReporte
      parameters:
        - in: query
          name: clienteId
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: from
          required: true
          schema:
            type: string
            format: date
        - in: query
          name: to
          required: true
          schema:
            type: string
            format: date
        - in: query
          name: format
          required: true
          schema:
            type: string
            enum: [json, pdf]
      responses:
        '200':
          description: Reporte generado
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ReportJson'
                  - $ref: '#/components/schemas/ReportPdf'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

components:
  parameters:
    PageParam:
      in: query
      name: page
      schema:
        type: integer
        minimum: 0
        default: 0
    SizeParam:
      in: query
      name: size
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 10
    ClienteIdParam:
      in: path
      name: clienteId
      required: true
      schema:
        type: string
        format: uuid
    CuentaIdParam:
      in: path
      name: cuentaId
      required: true
      schema:
        type: string
        format: uuid

  responses:
    BadRequest:
      description: Petición inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflicto
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  schemas:
    ErrorResponse:
      type: object
      required: [code, message, timestamp]
      properties:
        code:
          type: integer
        message:
          type: string
        details:
          type: string
        timestamp:
          type: string
          format: date-time

    CustomerDTO:
      type: object
      required: [id, name, gender, age, identification, address, phone, status]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        gender:
          type: string
        age:
          type: integer
        identification:
          type: string
        address:
          type: string
        phone:
          type: string
        password:
          type: string
        status:
          type: boolean

    CustomerCreate:
      type: object
      required: [name, gender, age, identification, address, phone, password, status]
      properties:
        name:
          type: string
        gender:
          type: string
        age:
          type: integer
        identification:
          type: string
        address:
          type: string
        phone:
          type: string
        password:
          type: string
        status:
          type: boolean

    CustomerUpdate:
      type: object
      required: [name, gender, age, identification, address, phone, status]
      properties:
        name:
          type: string
        gender:
          type: string
        age:
          type: integer
        identification:
          type: string
        address:
          type: string
        phone:
          type: string
        password:
          type: string
        status:
          type: boolean

    CustomerPatch:
      type: object
      properties:
        name:
          type: string
        gender:
          type: string
        age:
          type: integer
        identification:
          type: string
        address:
          type: string
        phone:
          type: string
        password:
          type: string
        status:
          type: boolean

    AccountDTO:
      type: object
      required: [id, accountNumber, accountType, initialBalance, status, customerId]
      properties:
        id:
          type: string
          format: uuid
        accountNumber:
          type: string
        accountType:
          type: string
          enum: [AHORRO, CORRIENTE]
        initialBalance:
          type: number
          format: double
        currentBalance:
          type: number
          format: double
        status:
          type: boolean
        customerId:
          type: string
          format: uuid

    AccountCreate:
      type: object
      required: [accountNumber, accountType, initialBalance, status, customerId]
      properties:
        accountNumber:
          type: string
        accountType:
          type: string
          enum: [AHORRO, CORRIENTE]
        initialBalance:
          type: number
          format: double
        status:
          type: boolean
        customerId:
          type: string
          format: uuid

    AccountPatch:
      type: object
      properties:
        accountType:
          type: string
          enum: [AHORRO, CORRIENTE]
        status:
          type: boolean

    TransactionDTO:
      type: object
      required: [id, date, transactionType, amount, balance, accountId]
      properties:
        id:
          type: string
          format: uuid
        date:
          type: string
          format: date-time
        transactionType:
          type: string
          enum: [CREDITO, DEBITO]
        amount:
          type: number
          format: double
        balance:
          type: number
          format: double
        accountId:
          type: string
          format: uuid

    TransactionCreate:
      type: object
      required: [accountId, transactionType, amount]
      properties:
        accountId:
          type: string
          format: uuid
        transactionType:
          type: string
          enum: [CREDITO, DEBITO]
        amount:
          type: number
          format: double

    ReportJson:
      type: object
      properties:
        customer:
          type: object
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
        range:
          type: object
          properties:
            from:
              type: string
              format: date
            to:
              type: string
              format: date
        accounts:
          type: array
          items:
            type: object
            properties:
              accountNumber:
                type: string
              accountType:
                type: string
              initialBalance:
                type: number
              status:
                type: boolean
              transactions:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    date:
                      type: string
                    transactionType:
                      type: string
                    amount:
                      type: number
                    balance:
                      type: number
                    accountId:
                      type: string
              totals:
                type: object
                properties:
                  debits:
                    type: number
                  credits:
                    type: number

    ReportPdf:
      type: object
      properties:
        base64:
          type: string

    PageResponse:
      type: object
      properties:
        content:
          type: array
          items:
            type: object
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer

    PageResponseCustomer:
      allOf:
        - $ref: '#/components/schemas/PageResponse'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/CustomerDTO'

    PageResponseAccount:
      allOf:
        - $ref: '#/components/schemas/PageResponse'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/AccountDTO'

    PageResponseTransaction:
      allOf:
        - $ref: '#/components/schemas/PageResponse'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/TransactionDTO'